requirements.txt
flask-sqlalchemy>=3.1.1
flask>=3.1.0
pytesseract>=0.3.13
trafilatura>=2.0.0
werkzeug>=3.1.3
psycopg2-binary>=2.9.10
email-validator>=2.2.0
gunicorn>=23.0.0
numpy>=2.2.1
pillow>=11.1.0
openai>=1.59.7
python-dotenv>=1.0.0
requests
beautifulsoup4